requirements.txt
fastapi[all]
motor
python-dotenv
sendgrid
python-jose[cryptography]
passlib[bcrypt]
